old-tests/submission/Infraware/Forms/contents/Forms/option_selected.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>option_selected</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2" selected> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var option = select.item(2);
		var selected = option.selected;
		option.selected = true;

		if (typeof(selected) == "boolean") {
			if (selected == false && option.selected == true) {
				test(function() {
					assert_true(true, "selected attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("selected attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("selected attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
